sql 변경에 따른 코드 변경 및 폴더명 변경

+++ src/main/java/com/d108/project/domain/logincredential/repository/LoginCredentialRepository.java
@@ -1,10 +1,10 @@
 package com.d108.project.domain.logincredential.repository;
 
-import com.d108.project.domain.logincredential.domain.LoginCredential;
+import com.d108.project.domain.logincredential.entity.LoginCredential;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoginCredentialRepository extends JpaRepository<LoginCredential, Integer> {
+public interface LoginCredentialRepository extends JpaRepository<LoginCredential, Long> {
     Optional<LoginCredential> findByUsername(String username);
 }
